+++ babel-preset/plugins/auto-importer.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015, Facebook, Inc.
+ * Copyright (c) 2013-present, Facebook, Inc.
  * All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
@@ -9,7 +9,7 @@
 
 'use strict';
 
-var MODULES = [
+const MODULES = [
   // Local Promise implementation.
   'Promise',
 ];
@@ -18,41 +18,41 @@ var MODULES = [
  * Automatically imports a module if its identifier is in the AST.
  */
 module.exports = function autoImporter(babel) {
-
-  var t = babel.types;
+  const t = babel.types;
 
   function isAppropriateModule(name, scope, state) {
-    var autoImported = state.get('autoImported');
+    const autoImported = state.autoImported;
     return MODULES.indexOf(name) !== -1
         && !autoImported.hasOwnProperty(name)
         && !scope.hasBinding(name, /*skip globals*/true);
   }
 
-  return new babel.Plugin('auto-importer', {
-    pre: function(state) {
+  return {
+    pre: function() {
       // Cache per file to avoid calling `scope.hasBinding` several
       // times for the same module, which has already been auto-imported.
-      state.set('autoImported', {});
+      this.autoImported = {};
     },
 
     visitor: {
-      ReferencedIdentifier: function(node, parent, scope, state) {
-        if (!isAppropriateModule(node.name, scope, state)) {
-          return node;
+      ReferencedIdentifier: function(path) {
+        const node = path.node;
+        const scope = path.scope;
+
+        if (!isAppropriateModule(node.name, scope, this)) {
+          return;
         }
 
         scope.getProgramParent().push({
           id: t.identifier(node.name),
           init: t.callExpression(
             t.identifier('require'),
-            [t.literal(node.name)]
+            [t.stringLiteral(node.name)]
           ),
         });
 
-        var autoImported = state.get('autoImported');
-        autoImported[node.name] = true;
-        state.set('autoImported', autoImported);
+        this.autoImported[node.name] = true;
       },
     },
-  });
+  };
 };