+++ gulpfile.js
@@ -1,20 +1,28 @@
-var gulp = require('gulp');
-var babel = require('gulp-babel');
-var flatten = require('gulp-flatten');
-var rename = require('gulp-rename');
-var del = require('del');
-var mergeStream = require('merge-stream');
-var runSequence = require('run-sequence');
+/**
+ * Copyright (c) 2013-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
 
-var babelDefaultOptions = require('fbjs-scripts/babel-6/default-options');
-var babelPluginDEV = require('fbjs-scripts/babel-6/dev-expression');
-var babelPluginModules = require('fbjs-scripts/babel-6/rewrite-modules');
-var gulpModuleMap = require('fbjs-scripts/gulp/module-map');
-var gulpStripProvidesModule = require('fbjs-scripts/gulp/strip-provides-module');
-var gulpCheckDependencies = require('fbjs-scripts/gulp/check-dependencies');
-var thirdPartyModuleMap = require('fbjs-scripts/third-party-module-map.json');
+'use strict';
 
-var paths = {
+const del = require('del');
+const flatten = require('gulp-flatten');
+const gulp = require('gulp');
+const gulpBabel = require('gulp-babel');
+const mergeStream = require('merge-stream');
+const rename = require('gulp-rename');
+const runSequence = require('run-sequence');
+
+const fbjsConfigurePreset = require('babel-preset-fbjs/configure');
+const gulpModuleMap = require('fbjs-scripts/gulp/module-map');
+const gulpStripProvidesModule = require('fbjs-scripts/gulp/strip-provides-module');
+const gulpCheckDependencies = require('fbjs-scripts/gulp/check-dependencies');
+
+const paths = {
   lib: {
     src: [
       'src/**/*.js',
@@ -22,19 +30,29 @@ var paths = {
       '!src/**/__mocks__/**/*.js',
     ],
     dest: 'lib',
+    presetOptions: {
+      stripDEV: true,
+      rewriteModules: {
+        map: require('fbjs-scripts/third-party-module-map'),
+      },
+    },
   },
   mocks: {
     src: [
       'src/**/__mocks__/**/*.js',
     ],
     dest: 'lib/__mocks__',
-    moduleOpts: {
-      prefix: '../',
+    presetOptions: {
+      stripDEV: true,
+      rewriteModules: {
+        map: require('fbjs-scripts/third-party-module-map'),
+        prefix: '../',
+      },
     },
   },
 };
 
-var moduleMapOpts = {
+const rewriteOptions = {
   moduleMapFile: './module-map.json',
   prefix: 'fbjs/lib/',
 };
@@ -44,22 +62,17 @@ gulp.task('clean', function() {
 });
 
 gulp.task('lib', function() {
-  var libTask = gulp
+  const libTask = gulp
     .src(paths.lib.src)
-    .pipe(gulpModuleMap(moduleMapOpts))
+    .pipe(gulpModuleMap(rewriteOptions))
     .pipe(gulpStripProvidesModule())
-    .pipe(babel(babelDefaultOptions({
-      plugins: [babelPluginDEV],
-    })))
+    .pipe(gulpBabel(fbjsConfigurePreset(paths.lib.presetOptions)))
     .pipe(flatten())
     .pipe(gulp.dest(paths.lib.dest));
 
-  var mockTask = gulp
+  const mockTask = gulp
     .src(paths.mocks.src)
-    .pipe(babel(babelDefaultOptions({
-      plugins: [babelPluginDEV],
-      moduleOpts: paths.mocks.moduleOpts,
-    })))
+    .pipe(gulpBabel(fbjsConfigurePreset(paths.mocks.presetOptions)))
     .pipe(flatten())
     .pipe(gulp.dest(paths.mocks.dest));
 
@@ -69,14 +82,17 @@ gulp.task('lib', function() {
 gulp.task('flow', function() {
   return gulp
     .src(paths.lib.src)
-    .pipe(gulpModuleMap(moduleMapOpts))
-    .pipe(babel({
-      plugins: [
-        "syntax-flow",
-        "syntax-trailing-function-commas",
-        "syntax-object-rest-spread",
-        [babelPluginModules, {map: thirdPartyModuleMap}]
-      ]
+    .pipe(gulpModuleMap(rewriteOptions))
+    .pipe(gulpBabel({
+      presets: [
+        fbjsConfigurePreset({
+          autoImport: false,
+          target: 'flow',
+          rewriteModules: {
+            map: require('fbjs-scripts/third-party-module-map'),
+          },
+        }),
+      ],
     }))
     .pipe(flatten())
     .pipe(rename({extname: '.js.flow'}))