+++ scripts/babel/default-options.js
@@ -9,45 +9,17 @@
 
 'use strict';
 
-var babelPluginModules = require('./rewrite-modules');
-var babelPluginAutoImporter = require('./auto-importer');
-var inlineRequires = require('./inline-requires');
+const assign = require('object-assign');
 
-var plugins = [
-  {
-    position: 'before',
-    transformer: babelPluginAutoImporter,
-  },
-  {
-    position: 'before',
-    transformer: babelPluginModules,
-  },
-];
-
-if (process.env.NODE_ENV === 'test') {
-  plugins.push({
-    position: 'after',
-    transformer: inlineRequires,
-  });
-}
-
-module.exports = {
-  nonStandard: true,
-  blacklist: [
-    'spec.functionName',
-  ],
-  loose: [
-    'es6.classes',
-  ],
-  stage: 1,
-  plugins: plugins,
-  _moduleMap: {
-    'core-js/library/es6/map': 'core-js/library/es6/map',
-    'core-js/library/es6/set': 'core-js/library/es6/set',
-    'isomorphic-fetch': 'isomorphic-fetch',
-    'promise': 'promise',
-    'promise/setimmediate/done': 'promise/setimmediate/done',
-    'promise/setimmediate/es6-extensions': 'promise/setimmediate/es6-extensions',
-    'ua-parser-js': 'ua-parser-js',
-  },
+module.exports = function(options) {
+  return {
+    presets: [
+      require('babel-preset-fbjs/configure')({
+        rewriteModules: assign({
+          map: require('../third-party-module-map'),
+        }, options.moduleOpts),
+      }),
+    ],
+    plugins: options.plugins || [],
+  };
 };